# Conflicts:
#	Modules/Test/classes/Setup/class.ilTest9DBUpdateSteps.php
#	Modules/Test/classes/class.ilObjTest.php

+++ Services/MediaObjects/classes/class.ilMediaAnalyzer.php
@@ -55,7 +55,7 @@ class ilMediaAnalyzer
 
     public function getPlaytimeSeconds(): int
     {
-        return $this->file_info["playtime_seconds"] ?? 0;
+        return (int) ($this->file_info["playtime_seconds"] ?? 0);
     }
 
     /**
